+++ src/import/converters.rs
@@ -1068,6 +1068,7 @@ mod tests {
                     None
                 },
                 masks_properties: None,
+                effects: None,
                 rotate_to_match_anim_pos_path: None,
                 matte_layer_index: None,
                 has_mask: None,